dev8

+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -1,34 +1,34 @@
 create table products (
     id          bigserial primary key,
     title       varchar(255),
-    price       int
+    price       numeric(8,2)
 );
 
 insert into products (title, price) values
-('Bread', 32),
-('Milk', 120),
-('Butter', 320),
-('Cheese', 500),
-('BreadTest', 400),
-('TestBread', 400),
-('Bread1', 32),
-('Milk1', 120),
-('Butter1', 320),
-('Cheese1', 500),
-('BreadTest1', 400),
-('TestBread1', 400),
-('Bread2', 32),
-('Milk2', 120),
-('Butter2', 320),
-('Cheese2', 500),
-('BreadTest2', 400),
-('TestBread2', 400);
+('Bread', 32.00),
+('Milk', 120.00),
+('Butter', 320.00),
+('Cheese', 500.00),
+('BreadTest', 400.00),
+('TestBread', 400.00),
+('Bread1', 32.00),
+('Milk1', 120.00),
+('Butter1', 320.00),
+('Cheese1', 500.00),
+('BreadTest1', 400.00),
+('TestBread1', 400.00),
+('Bread2', 32.00),
+('Milk2', 120.00),
+('Butter2', 320.00),
+('Cheese2', 500.00),
+('BreadTest2', 400.00),
+('TestBread2', 400.00);
 
 create table orders
 (
     id bigserial primary key,
     username varchar(255) not null,
-    total_price int not null,
+    total_price numeric(8,2) not null,
     address varchar(255),
     phone varchar(255),
     created_at timestamp default current_timestamp,
@@ -42,7 +42,7 @@ create table order_items
     order_id bigint not null references orders (id),
     quantity int not null,
     price_per_product int not null,
-    price int not null,
+    price numeric(8,2) not null,
     created_at timestamp default current_timestamp,
     updated_at timestamp default current_timestamp
 );
